+++ wikis/wikis/__init__.py
@@ -1,32 +1,22 @@
 from flask import (
-    Blueprint, render_template, request
+    Blueprint, render_template
 )
 
-from io import StringIO
-
-import csv2wiki
-
 bp = Blueprint('wikis', __name__, url_prefix='/wikis/', template_folder='templates')
+
+from wikis import create
 
 @bp.route('')
 def index():
     return render_template("upload.html");
 
-@bp.route('uploadcsv', methods=(["POST"]))
-def upload():
-    config = csv2wiki.parse_config_string(request.files["config"].read().decode("utf-8"))
-    csv_in = csv2wiki.CSVInput(StringIO(request.files["csv"].read().decode('utf-8')), config)
-    output = StringIO()
-    wiki_sess = csv2wiki.WikiSession(config, csv_in, False, output)
-    wiki_sess.make_pages(None, "size")
-    return "<pre>" + output.getvalue() + "</pre>"
-
 def main_partial():
     return "main_partial.html"
 
 # This section for standalone development/testing only
-if __name__ == "wikis":
+if __name__ == "__main__" or __name__ == "wikis":
     from flask import Flask
     app = Flask(__name__)
     app.register_blueprint(bp)
-    app.run(host='0.0.0.0')
+    if __name__ == "__main__":
+        app.run(host='0.0.0.0')
